Moved mcp install to utils
