+++ trymito.io/pages/install/vs-code.tsx
@@ -11,7 +11,7 @@ import {
   MCP_NAME,
   InstallMethod,
   McpConfig,
-} from "./mcpInstallConfig";
+} from "../../utils/mcpInstallConfig";
 
 const getVSCodeInstallUrl = (config: McpConfig): string => {
   const encodedConfig = encodeURIComponent(JSON.stringify(config));
