+++ trymito.io/utils/mcpInstallConfig.ts
